// src/server/shared/lib/document-handler/generated/models/ChunkSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { schemas__extract_schema__ChunkMetadataSchema } from './schemas__extract_schema__ChunkMetadataSchema';
export type ChunkSchema = {
  content: string;
  metadata: schemas__extract_schema__ChunkMetadataSchema;
};